pyproject.toml
[tool.poetry]
name = "sphinx.lionx.com.br"
version = "0.1.0"
description = ""
authors = ["Marco Sievers <im.ximit@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
py3dns = "^3.2.1"
fastapi-cache = "^0.1.0"
pip = "^21.1.2"
install = "^1.3.4"
validate_email = "^1.3"
pydantic = "^1.8.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
